+++ packages/base-driver/package.json
@@ -23,24 +23,21 @@
   },
   "license": "Apache-2.0",
   "author": "https://github.com/appium",
+  "main": "./build/lib/index.js",
   "types": "./build/lib/index.d.ts",
   "directories": {
     "lib": "lib"
   },
   "files": [
-    "index.js",
-    "index.d.ts",
     "lib",
+    "build/lib",
     "test-fixtures",
-    "build",
-    "tsconfig.json",
-    "!build/tsconfig.tsbuildinfo",
-    "!build/test"
+    "tsconfig.json"
   ],
   "scripts": {
     "test": "run-p test:unit test:types",
     "test:e2e": "mocha --exit --timeout 20s --slow 10s \"./test/e2e/**/*.spec.ts\"",
-    "test:smoke": "node ./index.js",
+    "test:smoke": "node ./build/lib/index.js",
     "test:unit": "mocha \"./test/unit/**/*.spec.ts\"",
     "test:types": "tsd"
   },
