+++ packages/base-driver/test/e2e/basedriver/driver.e2e.spec.ts
@@ -1,10 +1,10 @@
 import chai, {expect} from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 import type {Constraints, DriverCaps} from '@appium/types';
-import {BaseDriver, server, routeConfiguringFunction} from '../../../lib';
+import {server, routeConfiguringFunction} from '../../../lib';
 import {FakeDriver} from '../protocol/fake-driver';
 import axios from 'axios';
-import {TEST_HOST, getTestPort, driverE2ETestSuite} from '@appium/driver-test-support';
+import {TEST_HOST, getTestPort} from '@appium/driver-test-support';
 
 chai.use(chaiAsPromised);
 
@@ -12,12 +12,6 @@ const DEFAULT_CAPS = {
   platformName: 'fake',
   'appium:automationNAme': 'fake',
 };
-
-// @ts-expect-error BaseDriver constructor opts differ from DriverClass expectation
-driverE2ETestSuite(BaseDriver, {
-  platformName: 'iOS',
-  'appium:deviceName': 'Delorean',
-});
 
 describe('BaseDriver', function () {
   let port: number;
