+++ packages/base-driver/test/unit/basedriver/driver.spec.ts
@@ -1,12 +1,6 @@
 import {expect} from 'chai';
 import type {InitialOpts} from '@appium/types';
 import {BaseDriver} from '../../../lib';
-import {driverUnitTestSuite} from '@appium/driver-test-support';
-
-driverUnitTestSuite(BaseDriver as any, {
-  platformName: 'iOS',
-  'appium:deviceName': 'Delorean',
-});
 
 describe('BaseDriver', function () {
   describe('constructor', function () {
