+++ packages/driver-test-support/lib/index.ts
@@ -1,4 +1,1 @@
-export {createSessionHelpers, driverE2ETestSuite} from './e2e-suite';
-export * from './unit-suite';
-export * from './helpers';
-export type {NewSessionData, NewSessionResponse, SessionHelpers} from './types';
+export {TEST_HOST, getTestPort, createAppiumURL} from './helpers';
