+++ packages/driver-test-support/tsconfig.json
@@ -1,15 +1,20 @@
 {
   "extends": "@appium/tsconfig/tsconfig.json",
+  "ts-node": {
+    "transpileOnly": true,
+    "compilerOptions": {
+      "rootDir": "."
+    }
+  },
   "compilerOptions": {
     "strict": true,
     "rootDir": ".",
     "outDir": "build",
     "paths": {
-      "@appium/types": ["../types"],
-      "appium/driver": ["../base-driver"]
+      "@appium/types": ["../types"]
     },
-    "types": ["mocha", "chai", "chai-as-promised", "node"]
+    "types": ["mocha", "chai", "node"]
   },
   "include": ["lib", "test"],
-  "references": [{"path": "../types"}, {"path": "../base-driver"}]
+  "references": [{"path": "../types"}]
 }
