+++ packages/fake-driver/test/e2e/element-interaction-tests.ts
@@ -4,12 +4,12 @@ import {initSession, deleteSession, W3C_PREFIXED_CAPS} from '../helpers';
 
 chai.use(chaiAsPromised);
 
-export function elementTests() {
+export function elementTests(context: {port: number}) {
   describe('element interaction and introspection', function () {
     let driver: Awaited<ReturnType<typeof initSession>>;
 
     before(async function () {
-      driver = await initSession(W3C_PREFIXED_CAPS);
+      driver = await initSession(W3C_PREFIXED_CAPS, {port: context.port});
     });
     after(async function () {
       return await deleteSession(driver);
