+++ packages/fake-driver/test/e2e/find-element-tests.ts
@@ -4,12 +4,12 @@ import {initSession, deleteSession, W3C_PREFIXED_CAPS} from '../helpers';
 
 chai.use(chaiAsPromised);
 
-export function findElementTests() {
+export function findElementTests(context: {port: number}) {
   describe('finding elements', function () {
     let driver: Awaited<ReturnType<typeof initSession>>;
 
     before(async function () {
-      driver = await initSession(W3C_PREFIXED_CAPS);
+      driver = await initSession(W3C_PREFIXED_CAPS, {port: context.port});
     });
     after(async function () {
       return await deleteSession(driver);
