+++ packages/fake-driver/tsconfig.json
@@ -1,7 +1,10 @@
 {
   "extends": "@appium/tsconfig/tsconfig.json",
   "ts-node": {
-    "transpileOnly": true
+    "transpileOnly": true,
+    "compilerOptions": {
+      "rootDir": "."
+    }
   },
   "compilerOptions": {
     "rootDir": ".",
